Rename util.types into util.type_aliases
Before this commit, there was some confusion between the `type` library
and `raiden_contracts.utils.type`. Because of that, `join-contracts.py`
was not usable.

```
$ raiden_contracts/utils/join-contracts.py --import-map {}
raiden_contracts/contracts/TokenNetwork.sol flat.sol
Traceback (most recent call last):
  File "raiden_contracts/utils/join-contracts.py", line 2, in <module>
    import json
  File "/usr/lib/python3.6/json/__init__.py", line 106, in <module>
    from .decoder import JSONDecoder, JSONDecodeError
  File "/usr/lib/python3.6/json/decoder.py", line 3, in <module>
    import re
  File "/home/yh/src/raiden-contracts/venv/lib/python3.6/re.py", line
122, in <module>
    import enum
  File "/home/yh/src/raiden-contracts/venv/lib/python3.6/enum.py", line
2, in <module>
    from types import MappingProxyType, DynamicClassAttribute
  File "/home/yh/src/raiden-contracts/raiden_contracts/utils/types.py",
line 1, in <module>
    from typing import NewType
  File "/usr/lib/python3.6/typing.py", line 4, in <module>
    import contextlib
  File "/usr/lib/python3.6/contextlib.py", line 6, in <module>
    from functools import wraps
  File "/home/yh/src/raiden-contracts/venv/lib/python3.6/functools.py",
line 22, in <module>
    from types import MappingProxyType
ImportError: cannot import name 'MappingProxyType'
```
