+++ raiden_contracts/utils/__init__.py
@@ -6,5 +6,5 @@ from .merkle import *
 from .pending_transfers import *
 from .proofs import *
 from .signature import *
-from .types import *
+from .type_aliases import *
 from .bytecode import *
